Added Project Files for Delta Studios
Initial files for the "Pre-Alpha Version" of Delta Studios.
Delta Studios is  a game-and-software development project I am currently working on in the hopes of going professional at some point.

I am aware that some files / folders may not be in the right places at this time. Please feel free to advise me on how to do this properly, among anything else you can think of.