Git-ing Serious about Future-proofing Project
Yes, I'm aware that thats a terrible pun.

On a more serious note, however, I have decided to re-try implementing more bootstrap-navbar-elements into the header of my Project.
The "work-in-progress" part of this move is because everything - as far as I can tell anyway - is set to default-given values.
Additionally, I'm planning on tinkering with values to see what happens at a later time (once i finish focusing on some minor issues first) - hence, also, why this is in the Test branch.